Time: 2 ms (15.91%) | Memory: 9.8 MB (12.81%) - LeetSync

+++ 118-pascals-triangle/pascals-triangle.cpp
@@ -9,18 +9,20 @@ public:
         if (list.empty()) {
             temp.push_back(1);
             list.push_back(temp);
-        } else {
+        } 
+        else {
             if (list[i - 1].size() > 1) {
                 temp.push_back(1);
                 
-                for (size_t k = 0; k < list[i - 1].size() - 1; k++) {
+                for (int k = 0; k < list[i - 1].size() - 1; k++) {
                     int val = list[i - 1][k] + list[i - 1][k + 1];
                     temp.push_back(val);
                 }
                 
                 temp.push_back(1);
                 list.push_back(temp);
-            } else {
+            }
+            else {
                 temp.push_back(1);
                 temp.push_back(1);
                 list.push_back(temp);
